+++ src/App.tsx
@@ -1,36 +1,40 @@
 import 'bulma/css/bulma.css';
 import '@fortawesome/fontawesome-free/css/all.css';
 import './App.scss';
-import { Outlet, NavLink } from 'react-router-dom';
-import classNames from 'classnames';
+import { NavLink, Outlet, useLocation } from 'react-router-dom';
 
-const classNameLink = ({ isActive }: { isActive: boolean }) =>
-  classNames('navbar-item', {
-    'is-active': isActive,
-  });
+export const App = () => {
+  const location = useLocation();
 
-export const App = () => (
-  <>
-    <nav
-      className="navbar is-light is-fixed-top is-mobile has-shadow"
-      data-cy="Nav"
-    >
-      <div className="container">
-        <div className="navbar-brand">
-          <NavLink to="/" end className={classNameLink}>
-            Home
-          </NavLink>
-          <NavLink to="/tabs" className={classNameLink}>
-            Tabs
-          </NavLink>
+  return (
+    <>
+      <nav
+        className="navbar is-light is-fixed-top is-mobile has-shadow"
+        data-cy="Nav"
+      >
+        <div className="container">
+          <div className="navbar-brand">
+            <NavLink
+              to="/"
+              className={`navbar-item ${location.pathname === '/' ? 'is-active' : ''}`}
+            >
+              Home
+            </NavLink>
+            <NavLink
+              to="/tabs"
+              className={`navbar-item ${location.pathname.startsWith('/tabs') ? 'is-active' : ''}`}
+            >
+              Tabs
+            </NavLink>
+          </div>
+        </div>
+      </nav>
+
+      <div className="section">
+        <div className="container">
+          <Outlet />
         </div>
       </div>
-    </nav>
-
-    <div className="section">
-      <div className="container">
-        <Outlet />
-      </div>
-    </div>
-  </>
-);
+    </>
+  );
+};
