--- src/components/PageNotFound/PageNotFound.tsx
@@ -1,3 +1,0 @@
-export const PageNotFound = () => {
-  return <h1 className="title">Page not found</h1>;
-};
