--- src/components/PageNotFound/index.tsx
@@ -1,1 +1,0 @@
-export * from './PageNotFound';
