+++ src/components/TabsPage.tsx
@@ -1,0 +1,36 @@
+import classNames from 'classnames';
+import { NavLink, useParams } from 'react-router-dom';
+
+export const TabsPage = () => {
+  const tabs = [
+    { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
+    { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
+    { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
+  ];
+
+  const { tabId } = useParams<{ tabId?: string }>();
+
+  const activeTab = tabs.find(tab => tab.id === tabId);
+
+  return (
+    <>
+      <h1 className="title">Tabs page</h1>
+      <div className="tabs is-boxed">
+        <ul>
+          {tabs.map(t => (
+            <li
+              key={t.id}
+              data-cy="Tab"
+              className={classNames({ 'is-active': t.id === tabId })}
+            >
+              <NavLink to={`/tabs/${t.id}`}>{t.title}</NavLink>
+            </li>
+          ))}
+        </ul>
+      </div>
+      <div className="block" data-cy="TabContent">
+        {activeTab ? activeTab.content : 'Please select a tab'}
+      </div>
+    </>
+  );
+};
