--- src/components/TabsPage/index.tsx
@@ -1,1 +1,0 @@
-export * from './TabsPage';
